+++ Tests/FO-DICOM.Tests.Windows/Fixture.cs
@@ -1,0 +1,24 @@
+﻿using Xunit.Abstractions;
+using Xunit.Sdk;
+using FellowOakDicom;
+using System;
+
+namespace FellowOakDicom.Tests
+{
+    public sealed class InitializationFixture : IDisposable
+    {
+        public InitializationFixture()
+        {
+            new DicomSetupBuilder()
+                .RegisterServices(services =>
+                    services.AddWinFormsImaging()
+                )
+                .Build();
+        }
+
+        public void Dispose()
+        { }
+    }
+
+
+}